Allow runSubprocess to raise error instead of exiting
Allow runSubprocess to raise an error instead of running sys.exit(1). This
prevents the program from hanging if a subprocess killed while being run as
part of a multi-processing pool.
Remove the redundant function runMultiNoProgress, which can be handled by using
runMultiProgress with progress=False.
Switch iprscan-local.py to use runMultiProgress from funannotate.library,
instead of a redundant local copy of the function.